Initial commit of slackmech (v0.1.0)
This change introduces the first commit of `slackmech`. This package is built to
help mechanize API actions that Slack only permits the Web/Electron client to
make. So far, the functionality provided by this package allows you to log in to
Slack and to obtain a Session API Token. This token can be used with the
`github.comnlopes/slack` Client to make some requests.

Signed-off-by: Tim Heckman <t@heckman.io>