Add support for continuous integration using TravisCI
This adds support for Continuous Integration using Travis CI. This includes a
`Makefile` and a `.travis.yml` file.

This also updates the getSessionDetails test to pass on Linux, by making it more
generic. I was asserting too much of an error message in the test, and it
included a darwin OS specific-string that would never assert to true on Linux.

Signed-off-by: Tim Heckman <t@heckman.io>