Install golangci-lint and use it in tests
This updates the `Makefile` to include a build target that installs
`golangci-lint` and one to run it. The `test` build target also now runs the
linter by default.

This will fail the build on `master` at the time of commit, and will be used to
test an incoming PR's linter fixes.

Signed-off-by: Tim Heckman <t@heckman.io>