Add ability to view and manage pending Slack invitation
This change introduces the ability to view the current pending invites, as well
as to resend and revoke them. The need for this arose after I found that the
Gophers Slack Workspace had almost 2,300 pending invitations. This caused that
page to take forever to load, and to subsequently make it harder to help users
who were having issues.

Signed-off-by: Tim Heckman <t@heckman.io>